third_party/WebKit/LayoutTests/imported/web-platform-tests/FileAPI/blob/Blob-XHR-revoke-expected.txt

This is a testharness.js-based test.
FAIL Revoking blob URL used with XMLHttpRequest assert_equals: expected "test" but got ""
Harness: the test ran to completion.
